Quote-Generation/backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
python-dotenv
databricks-sql-connector